app/oath2.py
from jose import JWSError, jwt
from datetime import datetime, timedelta


SECRET_KEY = "94e8009e01b9961deed532ef826577808ed260a4bb58dba57a1ab0020e668a6c"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

    return encoded_jwt
