pyproject.toml
[tool.poetry]
name = "10k_day-api"
version = "0.1.0"
description = ""
authors = ["Dominik Piekarczyk <piekarczyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
pytest = "^6.2.5"
requests = "^2.26.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.27"
black = "^21.11b1"
psycopg2 = "^2.9.2"
email-validator = "^1.1.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
routers = "^0.10.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
